Added an implementation for test_delete_saved_weather_by_city_returns_message_when_weather_not_in_db

+++ src/weather/weather.py
@@ -15,6 +15,8 @@ class WeatherApp(object):
         return self.__weather_database
 
     def delete_saved_weather_by_city(self, city_name):
+        if self.__weather_database.find(city_name) == []:
+            return f"{city_name} not in database"
         self.__weather_database.delete(city_name)
 
     def save_weather_single_city(self, city_name):
